+++ src/haddock/restraints/passive_from_active.py
@@ -15,6 +15,8 @@ import numpy as np
 import sys
 from Bio.PDB import PDBParser, NeighborSearch
 
+from haddock.libs.librestraints import get_surface_resids
+
 
 def add_pass_from_act_arguments(pass_from_act_subcommand):
     """Add arguments to the pass_from_act subcommand."""
@@ -46,135 +48,6 @@ def add_pass_from_act_arguments(pass_from_act_subcommand):
         )
 
     return pass_from_act_subcommand
-
-# Scaling factors for relative ASA
-# Calculated using extended ALA-X-ALA peptides
-# Taken from NACCESS
-rel_asa = {
-    'total':
-        {
-            'ALA': 107.95,
-            'CYS': 134.28,
-            'ASP': 140.39,
-            'GLU': 172.25,
-            'PHE': 199.48,
-            'GLY': 80.10,
-            'HIS': 182.88,
-            'ILE': 175.12,
-            'LYS': 200.81,
-            'LEU': 178.63,
-            'MET': 194.15,
-            'ASN': 143.94,
-            'PRO': 136.13,
-            'GLN': 178.50,
-            'ARG': 238.76,
-            'SER': 116.50,
-            'THR': 139.27,
-            'VAL': 151.44,
-            'TRP': 249.36,
-            'TYR': 212.76,
-        },
-    'bb':
-        {
-            'ALA': 38.54,
-            'CYS': 37.53,
-            'ASP': 37.70,
-            'GLU': 37.51,
-            'PHE': 35.37,
-            'GLY': 47.77,
-            'HIS': 35.80,
-            'ILE': 37.16,
-            'LYS': 37.51,
-            'LEU': 37.51,
-            'MET': 37.51,
-            'ASN': 37.70,
-            'PRO': 16.23,
-            'GLN': 37.51,
-            'ARG': 37.51,
-            'SER': 38.40,
-            'THR': 37.57,
-            'VAL': 37.16,
-            'TRP': 38.10,
-            'TYR': 35.38,
-        },
-    'sc':
-        {
-            'ALA': 69.41,
-            'CYS': 96.75,
-            'ASP': 102.69,
-            'GLU': 134.74,
-            'PHE': 164.11,
-            'GLY': 32.33,
-            'HIS': 147.08,
-            'ILE': 137.96,
-            'LYS': 163.30,
-            'LEU': 141.12,
-            'MET': 156.64,
-            'ASN': 106.24,
-            'PRO': 119.90,
-            'GLN': 140.99,
-            'ARG': 201.25,
-            'SER': 78.11,
-            'THR': 101.70,
-            'VAL': 114.28,
-            'TRP': 211.26,
-            'TYR': 177.38,
-        }
-}
-
-def get_surface_resids(structure, cutoff=15):
-    """
-    Calls freesasa using its Python API and returns
-    per-residue accessibilities.
-    """
-    try:
-        from freesasa import Classifier, structureFromBioPDB, calc
-    except ImportError as err:
-        print('[!] The binding affinity prediction tools require the \'freesasa\' Python API', file=sys.stderr)
-        raise ImportError(err)
-
-    asa_data, rsa_data, rel_main_chain, rel_side_chain = {}, {}, {}, {}
-    _rsa = rel_asa['total']
-    _rsa_bb = rel_asa['bb']
-    _rsa_sc = rel_asa['sc']
-
-    #classifier = Classifier(config_path)
-    classifier = Classifier()
-
-    struct = structureFromBioPDB(structure, classifier, )
-    result = calc(struct)
-
-    # iterate over all atoms to get SASA and residue name
-    for idx in range(struct.nAtoms()):
-        atname = struct.atomName(idx).strip()
-        resname = struct.residueName(idx)
-        resid = int(struct.residueNumber(idx))
-        chain = struct.chainLabel(idx)
-        at_uid = (chain, resname, resid, atname)
-        res_uid = (chain, resname, resid)
-
-        asa = result.atomArea(idx)
-        asa_data[at_uid] = asa
-        # add asa to residue
-        rsa_data[res_uid] = rsa_data.get(res_uid, 0) + asa
-
-        if atname in ('C', 'N', 'O'):
-            rel_main_chain[res_uid] = rel_main_chain.get(res_uid, 0) + asa
-        else:
-            rel_side_chain[res_uid] = rel_side_chain.get(res_uid, 0) + asa
-
-    # convert total asa ro relative asa
-    rsa_data.update((res_uid, asa / _rsa[res_uid[1]]) for res_uid, asa in rsa_data.items())
-    rel_main_chain.update((res_uid, asa / _rsa_bb[res_uid[1]] * 100) for res_uid, asa in rel_main_chain.items())
-    rel_side_chain.update((res_uid, asa / _rsa_sc[res_uid[1]] * 100) for res_uid, asa in rel_side_chain.items())
-
-    # We format to fit the pipeline
-    resid_access = {}
-    for res_uid, access in rel_main_chain.items():
-        resid_access[res_uid[2]] = {'side_chain_rel': rel_side_chain.get(res_uid), 'main_chain_rel': access}
-    surface_resids = [r for r, v in resid_access.items() if v['side_chain_rel'] >= cutoff or
-                      v['main_chain_rel'] >= cutoff]
-    return surface_resids
 
 
 def passive_from_active(structure, active_list, chain_id=None, surface_list=""):
