+++ tests/test_cli_restraints.py
@@ -117,4 +117,4 @@ def test_restrain_bodies_exclude(protdna_input_list, capsys):
     restrain_bodies(protdna_input_list[0].rel_path, exclude="A")
     captured = capsys.readouterr()
     out_lines = captured.out.split("\n")
-    assert out_lines[0] == "assign (segid B and resi 3 and name P) (segid B and resi 29 and name P) 31.170 0.0 0.0"
+    assert out_lines[0] == "assign (segid B and resi 6 and name P) (segid B and resi 35 and name P) 15.187 0.0 0.0"
